requirements.txt
numpy>=1.21.0
scikit-learn>=1.0.0
flask>=2.0.0
nltk>=3.8.1
language-tool-python>=2.9.4
python-docx>=0.8.11
PyPDF2>=3.0.0
python-dotenv>=0.19.0 
